+++ src/output.c
@@ -5,7 +5,7 @@
 #include <sys/time.h>
 #include <stdio.h>
 
-static void find_field(char c, field *fields, field_value *values, field **f, field_value **v) {
+static void find_field(char c, const field *fields, const field_value *values, const field **f, const field_value **v) {
     for(; fields->label; fields++, values++) 
         if(fields->c == c) {
             if(f) *f = fields;
@@ -14,19 +14,19 @@ static void find_field(char c, field *fields, field_value *values, field **f, fi
         }
 }
 
-static int is_field(char c, field *fields) {
+static int is_field(char c, const field *fields) {
     for(; fields->label; fields++) 
         if(fields->c == c) return 1;
     return 0;
 }
 
 
-static void render_cols(int count, field *fields, field_value *values, char *selector) {
+static void render_cols(int count, const field *fields, const field_value *values, const char *selector) {
     struct tm fmt;
     int first = 1;
     for(; *selector; selector++) {
-        field *f;
-        field_value *v;
+        const field *f;
+        const field_value *v;
         find_field(*selector, fields, values, &f, &v);
         if(!first)
             printf(" ");
@@ -45,6 +45,9 @@ static void render_cols(int count, field *fields, field_value *values, char *sel
             case fld_type_string:
                 printf("%s", v->value.string_value);
                 break;
+            case fld_type_int:
+                printf("%d", v->value.int_value);   
+                break;
         }
         first = 0;
     }
@@ -53,23 +56,23 @@ static void render_cols(int count, field *fields, field_value *values, char *sel
 
 static int width = 0;
 
-static void render_rows(int count, field *fields, field_value *values, char *selector) {
+static void render_rows(int count, const field *fields, const field_value *values, const char *selector) {
     if(count) {
         for(size_t l=0; l<width; l++) printf("-");
         printf("\n");
     }
 
     int max = 0;
-    for(char *s=selector; *s; s++) {
-        field *f;
+    for(const char *s=selector; *s; s++) {
+        const field *f;
         find_field(*s, fields, values, &f, NULL);
         if(strlen(f->label) > max) 
             max = strlen(f->label);
     }
 
-    for(char *s=selector; *s; s++) {
-        field *f;
-        field_value *v;
+    for(const char *s=selector; *s; s++) {
+        const field *f;
+        const field_value *v;
         find_field(*s, fields, values, &f, &v);
         printf("%*s : ", max, f->label);
         struct tm fmt;
@@ -89,21 +92,24 @@ static void render_rows(int count, field *fields, field_value *values, char *sel
             case fld_type_string:
                 printf("%s%n", v->value.string_value, &w);
                 break;
+            case fld_type_int:
+                printf("%d%n", v->value.int_value, &w);
+                break;
         }
         if(!count && max+w+3 > width) width=max+w+3;
         printf("\n");
     }
 }
 
-void render(int count, field *fields, field_value *values, char *selector, int rows) {
+void render(int count, const field *fields, const field_value *values, const char *selector, int rows) {
     if(rows) render_rows(count, fields, values, selector);
     else render_cols(count, fields, values, selector);
 }
 
-void render_headers(field *fields, char *selector) {
+void render_headers(const field *fields, const char *selector) {
     int first = 1;
     for(; *selector; selector++) {
-        field *f;
+        const field *f;
         find_field(*selector, fields, NULL, &f, NULL);
         if(!first) printf(" ");
         printf("%s", f->label_short);
@@ -113,7 +119,7 @@ void render_headers(field *fields, char *selector) {
 
 }
 
-int check_selector(field *fields, char *selector) {
+int check_selector(const field *fields, const char *selector) {
     for(; *selector; selector++)
         if(!is_field(*selector, fields)) return -1;