+++ src/output.h
@@ -10,7 +10,8 @@ typedef struct {
         fld_type_time_string,
         fld_type_time, /* Time as number of seconds since epoch */
         fld_type_double,
-        fld_type_string
+        fld_type_string,
+        fld_type_int
     } type;
 } field;
 
@@ -18,13 +19,14 @@ typedef struct {
     union {
         time_t time_value;
         double double_value;
-        char *string_value;
+        const char *string_value;
+        int int_value;
     } value;
 } field_value;
 
-int check_selector(field *fields, char *selector);
+int check_selector(const field *fields, const char *selector);
 
-void render(int count, field *fields, field_value *values, char *selector, int rows);
+void render(int count, const field *fields, const field_value *values, const char *selector, int rows);
 
-void render_headers(field *fields, char *selector);
+void render_headers(const field *fields, const char *selector);
 #endif