+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 Next
 ====
 * Add ECI X, Y and Z to `satpass` output fields
+* Make `tleinfo` output configurable with fields-string
 
 1.0.0
 =====